Recovery is working again. Blind descriptions are working now for weapons, pieces of armor, and tools.

+++ src/files.c
@@ -3681,7 +3681,7 @@ recover_savefile()
 {
     int gfd, lfd, sfd;
 	int lev, savelev, hpid;
-	size_t pltmpsiz;
+	int pltmpsiz;
     xchar levc;
     struct version_info version_data;
     int processed[256];
